ui: factor out global components
