+++ src/App.js
@@ -8,11 +8,11 @@ import 'bulma/css/bulma.css'
 import './App.css'
 import config from './data/config'
 import server from './data/server'
-import Navbar from './components/Navbar'
+import Navbar from './components/global/Navbar'
 import Home from './components/Home'
 import Login from './components/Login'
 import Dashboard from './components/account/Dashboard'
-import Footer from './components/Footer'
+import Footer from './components/global/Footer'
 import Send from './components/account/Send'
 import Receive from './components/account/Receive'
